renamed: display_menu to message_prompts
