+++ src/config/file_paths.py
@@ -1,0 +1,8 @@
+'''Contains all the file paths'''
+
+
+class FilePaths:
+    '''Contains all the file paths as class variables'''
+
+    DATABASE_PATH = 'src\\database\\data.db'
+    QUESTIONS_JSON_PATH = 'src\\config\\questions.json'
