+++ src/controllers/auth_controller.py
@@ -5,7 +5,7 @@ import logging
 import sqlite3
 from typing import Tuple
 
-from config.display_menu import DisplayMessage, Headers, LogMessage
+from config.message_prompts import DisplayMessage, Headers, LogMessage
 from config.queries import Queries
 from config.regex_patterns import RegexPattern
 from database.database_access import DatabaseAccess as DAO
