+++ src/controllers/quiz_controller.py
@@ -5,7 +5,7 @@ import sqlite3
 import time
 from typing import List, Tuple
 
-from config.display_menu import DisplayMessage, Headers, LogMessage
+from config.message_prompts import DisplayMessage, Headers, LogMessage
 from config.queries import Queries
 from config.regex_patterns import RegexPattern
 from controllers.helpers import quiz_helper as QuizHelper
