+++ src/controllers/user_controller.py
@@ -6,7 +6,7 @@ import sqlite3
 import string
 from typing import List, Tuple
 
-from config.display_menu import DisplayMessage, Headers, LogMessage
+from config.message_prompts import DisplayMessage, Headers, LogMessage
 from config.queries import Queries
 from config.regex_patterns import RegexPattern
 from database.database_access import DatabaseAccess as DAO
