+++ src/database/database_access.py
@@ -1,8 +1,10 @@
 '''Contains methods for establishing database connection'''
 
 from typing import List, Tuple
-from database.database_connection import DatabaseConnection
+
+from config.file_paths import FilePaths
 from config.queries import InitializationQueries
+from database.database_connection import DatabaseConnection
 
 
 class DatabaseAccess:
@@ -12,7 +14,7 @@ class DatabaseAccess:
     def read_from_database(query: str, data: Tuple = None) -> List:
         '''Reads data from database'''
 
-        with DatabaseConnection('src\\database\\data.db') as connection:
+        with DatabaseConnection(FilePaths.DATABASE_PATH) as connection:
             cursor = connection.cursor()
             if not data:
                 cursor.execute(query)
@@ -25,7 +27,7 @@ class DatabaseAccess:
     def write_to_database(query: str, data: Tuple = None) -> None:
         ''' CREATE TABLE / Add / Update / Delete data from database'''
 
-        with DatabaseConnection('src\\database\\data.db') as connection:
+        with DatabaseConnection(FilePaths.DATABASE_PATH) as connection:
             cursor = connection.cursor()
             cursor.execute(InitializationQueries.ENABLE_FOREIGN_KEYS)
             if not data:
